Pass along props to DOM elements
This technique gives more flexibility when using the components.

There is no more flickering when a card is clicked right at the end of
the game: instead of changing the card from a button to a div, the
button stays as a button and is instead disabled, allowing the "active"
background color to fade instead of disappearing abruptly.