Optimize Game component
There was a double re-render of the game component when a high score was
beaten: the component changed its state to go from the count asking
phase to the scoring phase which caused a re-render, and then it
registered a new high score which made the parent change its `highScore`
prop and caused another re-render.

We don't use the `highScore` prop when rendering so we can
just exclude it from `shouldComponentUpdate`.